# docs/openapi/swagger.yaml
definitions:
...
  api_rest_strudels.ConversationMessageDTO:
    properties:
      clarifying_questions:
        items:
          type: string
        type: array
      content:
        type: string
      created_at:
        type: string
      doc_references:
        items:
          $ref: '#/definitions/api_rest_strudels.DocReferenceDTO'
        type: array
      id:
        type: string
      is_actionable:
        type: boolean
      is_code_response:
        type: boolean
      role:
        type: string
      strudel_references:
        items:
          $ref: '#/definitions/api_rest_strudels.StrudelReferenceDTO'
        type: array
    type: object
...
  api_rest_strudels.MessageResponse:
    properties:
      message:
        type: string
    type: object
  api_rest_strudels.StrudelDetailResponse:
    properties:
      categories:
        items:
          type: string
        type: array
      cc_signal:
        $ref: '#/definitions/codeberg_org_algopatterns_server_algopatterns_strudels.CCSignal'
      code:
        type: string
      conversation_history:
        items:
          $ref: '#/definitions/api_rest_strudels.ConversationMessageDTO'
        type: array
      created_at:
        type: string
      description:
        type: string
      forked_from:
        type: string
      id:
        type: string
      is_public:
        type: boolean
      parent_cc_signal:
        $ref: '#/definitions/codeberg_org_algopatterns_server_algopatterns_strudels.CCSignal'
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  api_rest_strudels.StrudelsListResponse:
    properties:
      pagination:
        $ref: '#/definitions/codeberg_org_algopatterns_server_api_rest_pagination.Meta'
      strudels:
        items:
          $ref: '#/definitions/codeberg_org_algopatterns_server_algopatterns_strudels.Strudel'
        type: array
    type: object
  api_rest_strudels.TagsListResponse:
    properties:
      tags:
        items:
          type: string
        type: array
    type: object
...
  codeberg_org_algopatterns_server_algopatterns_strudels.CCSignal:
    enum:
    - cc-cr
    - cc-dc
    - cc-ec
    - cc-op
    - no-ai
    type: string
    x-enum-comments:
      CCSignalCredit: 'Credit: Allow AI use with attribution'
      CCSignalDirect: 'Credit + Direct: Attribution + financial/in-kind support'
      CCSignalEcosystem: 'Credit + Ecosystem: Attribution + contribute to commons'
      CCSignalNoAI: 'No AI: Explicitly opt-out of AI training'
      CCSignalOpen: 'Credit + Open: Attribution + keep derivatives open'
    x-enum-descriptions:
    - 'Credit: Allow AI use with attribution'
    - 'Credit + Direct: Attribution + financial/in-kind support'
    - 'Credit + Ecosystem: Attribution + contribute to commons'
    - 'Credit + Open: Attribution + keep derivatives open'
    - 'No AI: Explicitly opt-out of AI training'
    x-enum-varnames:
    - CCSignalCredit
    - CCSignalDirect
    - CCSignalEcosystem
    - CCSignalOpen
    - CCSignalNoAI
  codeberg_org_algopatterns_server_algopatterns_strudels.CreateStrudelRequest:
    properties:
      categories:
        description: max 10 categories, each max 50 chars
        items:
          type: string
        maxItems: 10
        type: array
      cc_signal:
        $ref: '#/definitions/codeberg_org_algopatterns_server_algopatterns_strudels.CCSignal'
      code:
        description: 1MB limit
        maxLength: 1048576
        type: string
      conversation_history:
        description: max 100 messages
        items:
          $ref: '#/definitions/codeberg_org_algopatterns_server_internal_agent.Message'
        maxItems: 100
        type: array
      description:
        maxLength: 2000
        type: string
      forked_from:
        type: string
      is_public:
        type: boolean
      license:
        type: string
      tags:
        description: max 20 tags, each max 50 chars
        items:
          type: string
        maxItems: 20
        type: array
      title:
        maxLength: 200
        type: string
    required:
    - code
    - title
    type: object
  codeberg_org_algopatterns_server_algopatterns_strudels.Strudel:
    properties:
      ai_assist_count:
        type: integer
      categories:
        items:
          type: string
        type: array
      cc_signal:
        $ref: '#/definitions/codeberg_org_algopatterns_server_algopatterns_strudels.CCSignal'
      code:
        type: string
      conversation_history:
        items:
          $ref: '#/definitions/codeberg_org_algopatterns_server_internal_agent.Message'
        type: array
      created_at:
        type: string
      description:
        type: string
      forked_from:
        type: string
      id:
        type: string
      is_public:
        type: boolean
      license:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  codeberg_org_algopatterns_server_algopatterns_strudels.UpdateStrudelRequest:
    properties:
      categories:
        items:
          type: string
        maxItems: 10
        type: array
      cc_signal:
        $ref: '#/definitions/codeberg_org_algopatterns_server_algopatterns_strudels.CCSignal'
      code:
        description: 1MB limit
        maxLength: 1048576
        type: string
      conversation_history:
        items:
          $ref: '#/definitions/codeberg_org_algopatterns_server_internal_agent.Message'
        maxItems: 100
        type: array
      description:
        maxLength: 2000
        type: string
      is_public:
        type: boolean
      license:
        type: string
      tags:
        items:
          type: string
        maxItems: 20
        type: array
      title:
        maxLength: 200
        type: string
    type: object
...
  codeberg_org_algopatterns_server_api_rest_pagination.Meta:
    properties:
      has_more:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  codeberg_org_algopatterns_server_internal_agent.Message:
    properties:
      content:
        description: message content
        type: string
      is_code_response:
        description: true if AI generated code
        type: boolean
      role:
        description: '"user" or "assistant"'
        type: string
    type: object
  codeberg_org_algopatterns_server_internal_attribution.StrudelStats:
    properties:
      fork_count:
        type: integer
      last_used_at:
        type: string
      total_uses:
        type: integer
      unique_users:
        type: integer
    type: object
  codeberg_org_algopatterns_server_internal_attribution.StrudelStatsResponse:
    properties:
      recent_uses:
        items:
          $ref: '#/definitions/codeberg_org_algopatterns_server_internal_attribution.StrudelUse'
        type: array
      stats:
        $ref: '#/definitions/codeberg_org_algopatterns_server_internal_attribution.StrudelStats'
    type: object
  codeberg_org_algopatterns_server_internal_attribution.StrudelUse:
    properties:
      created_at:
        type: string
      id:
        type: string
      requesting_display_name:
        type: string
      requesting_user_id:
        type: string
      similarity_score:
        type: number
      target_strudel_id:
        type: string
      target_strudel_title:
        type: string
    type: object
  codeberg_org_algopatterns_server_internal_errors.ErrorResponse:
    properties:
      details:
        description: optional details (sanitized in production)
        type: string
      error:
        description: error code (e.g., "unauthorized", "not_found")
        type: string
      message:
        description: user-friendly message
        type: string
      request_id:
        description: echoed from request for correlation
        type: string
    type: object
host: algopatterns.dev
info:
  contact:
    name: API Support
    url: https://codeberg.org/algopatterns/server
  description: |-
    AI-powered Strudel code generation and collaborative live coding platform

    Features:
    - AI-powered Strudel code generation from natural language
    - Real-time collaborative editing via WebSockets
    - OAuth authentication (Google, GitHub)
    - Anonymous session support
    - Save and share Strudel patterns
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
  title: Algopatterns API
  version: "1.0"
paths:
...
  /api/v1/public/strudels:
    get:
      description: Get publicly shared strudels from all users with pagination, search,
        and filtering
      parameters:
      - default: 20
        description: Items per page (max 100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of items to skip
        in: query
        name: offset
        type: integer
      - description: Search in title and description
        in: query
        name: search
        type: string
      - collectionFormat: csv
        description: Filter by tags (comma-separated)
        in: query
        items:
          type: string
        name: tags
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_rest_strudels.StrudelsListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
      summary: List public strudels
      tags:
      - strudels
  /api/v1/public/strudels/{id}:
    get:
      description: Get a publicly shared strudel by its ID (for forking)
      parameters:
      - description: Strudel ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_algopatterns_strudels.Strudel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
      summary: Get public strudel by ID
      tags:
      - strudels
  /api/v1/public/strudels/{id}/stats:
    get:
      description: Get attribution stats for a public strudel (how many times it was
        used as RAG context)
      parameters:
      - description: Strudel ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_attribution.StrudelStatsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
      summary: Get strudel usage stats
      tags:
      - strudels
  /api/v1/public/strudels/tags:
    get:
      description: Get all unique tags from public strudels
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_rest_strudels.TagsListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
      summary: List public tags
      tags:
      - strudels
...
  /api/v1/strudels:
    get:
      description: Get strudels owned by the authenticated user with pagination, search,
        and filtering
      parameters:
      - default: 20
        description: Items per page (max 100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of items to skip
        in: query
        name: offset
        type: integer
      - description: Search in title and description
        in: query
        name: search
        type: string
      - collectionFormat: csv
        description: Filter by tags (comma-separated)
        in: query
        items:
          type: string
        name: tags
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_rest_strudels.StrudelsListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List user's strudels
      tags:
      - strudels
    post:
      consumes:
      - application/json
      description: Save a new Strudel pattern with code, title, and metadata
      parameters:
      - description: Strudel data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/codeberg_org_algopatterns_server_algopatterns_strudels.CreateStrudelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_algopatterns_strudels.Strudel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create strudel
      tags:
      - strudels
  /api/v1/strudels/{id}:
    delete:
      description: Delete a strudel (must be owner)
      parameters:
      - description: Strudel ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_rest_strudels.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete strudel
      tags:
      - strudels
    get:
      description: Get a specific strudel by ID (owner or public)
      parameters:
      - description: Strudel ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_rest_strudels.StrudelDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
      summary: Get strudel by ID
      tags:
      - strudels
    put:
      consumes:
      - application/json
      description: Update a strudel's properties (must be owner)
      parameters:
      - description: Strudel ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/codeberg_org_algopatterns_server_algopatterns_strudels.UpdateStrudelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_algopatterns_strudels.Strudel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update strudel
      tags:
      - strudels
  /api/v1/strudels/tags:
    get:
      description: Get all unique tags from the authenticated user's strudels
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_rest_strudels.TagsListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List user's tags
      tags:
      - strudels
...
securityDefinitions:
  BearerAuth:
    description: 'JWT token for authenticated requests. Format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"